Update project setup with npm
Added the vite bundler/compiler so I could get autocomplete on p5js functions. Now each 'experiment'
is loaded dynamically, which removes the requirement to specify the sketch and its data separately on a JSON file.
Now new experiments can be simply written as JS modules in `src/experiments`, which must export a sketch function and
a name and description.